requirements.txt
streamlit
langchain
langchain-core
langchain-groq
python-dotenv
groq
pydantic
